requirements.txt
annotated-types==0.7.0
pydantic==2.10.1
pydantic_core==2.27.1
PyYAML==6.0.2
typing_extensions==4.12.2
